[coq] [wip] Allow Coq libraries to depend on ML plugins.
We add a new field `(libraries <ocaml_libraries>)` to the `coqlib`
stanza for specifying dependencies on OCaml libs, such as plugins.

The current implementation has to workaround two `coqdep` problems:

- `coqdep` depends on `.cmo` files by default, however Dune produces
  `cma`;
- `-I` flags to `coqdep` must refer to the source tree as to find the
  `.cmxs` file.

I am not sure if `coqdep` needs the `.mlpack` file to be present in
the plugin source tree. In any case this should be fixed upstream.

We should improve `coqdep` so this hacks are not necessary.

TODO: current rules don't correctly call `coqdep` with the sources of
plugin set, thus the rules are properly depending on the correct ML
source artifacts.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>